Update UI with dynamic filtering and modify database schema
The commit introduces dynamic filtering options to the user interface for better testing. It also includes updates to the SQL database schema with new fields and changes to table structures. Several classes have been relocated and renamed to improve organization. Specific enhancements were made to the 'dto_mapper' classes and the 'MissionCache' class for better functionality and readability.

+++ src/main/java/com/db_course/entity_model/AtmosphereElement.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,6 +14,6 @@ public class AtmosphereElement {
 
     private int atmosphereId;
     private int elementId;
-    private double percentage;
+    private BigDecimal percentage;
 
 }
